requirements.txt
pandas
PyPDF2
yagmail
python-dotenv
openpyxl
PyMuPDF
reportlab 